+++ FitNesseRoot/DbFit/AcceptanceTests/JavaTests/HsqlTests/FlowMode/SetUp/content.txt
@@ -9,3 +9,7 @@
 |Execute Ddl|${CREATEMULTIPLYPROC}|
 
 |Execute Ddl|${CREATECALCLENGTHPROC}|
+
+|Execute Ddl|${CREATERAISEERRORNOPARAMSPROC}|
+
+|Execute Ddl|${CREATERAISEERRORWITHPARAMSPROC}|
