+++ FitNesseRoot/DbFit/AcceptanceTests/JavaTests/HsqlTests/StandaloneFixtures/SetUp/content.txt
@@ -11,3 +11,7 @@
 |Execute Ddl|${CREATEMULTIPLYPROC}|
 
 |Execute Ddl|${CREATECALCLENGTHPROC}|
+
+|Execute Ddl|${CREATERAISEERRORNOPARAMSPROC}|
+
+|Execute Ddl|${CREATERAISEERRORWITHPARAMSPROC}|
