+++ dbfit-java/hsqldb/src/main/java/dbfit/environment/HSQLDBEnvironment.java
@@ -279,5 +279,12 @@ public class HSQLDBEnvironment extends AbstractDbEnvironment {
     public StatementExecution createFunctionStatementExecution(PreparedStatement statement) {
         return new StatementExecutionCapturingResultSetValue(statement);
     }
-}
 
+    public String getActualErrorCode(SQLException e) {
+        if (e.getSQLState().equals("5800")) { // unhandled user-defined exception 
+            return e.getNextException().getSQLState();        	
+        } else {
+        	return e.getSQLState();
+        }
+    }
+}
