+++ dbfit-java/mysql/src/integration-test/resources/001_add_objects_needed_for_acceptance_test.sql
@@ -12,4 +12,6 @@ create procedure createuser(IN newname varchar(100), IN newusername varchar(100)
 
 create procedure Multiply(IN factor int, INOUT val int) set val =val*factor;
 
-CREATE PROCEDURE raise_error_no_params() BEGIN SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'test exception'; END;
+CREATE PROCEDURE raise_error_no_params() BEGIN SIGNAL SQLSTATE '20001' SET MESSAGE_TEXT = 'test exception'; END;
+
+CREATE PROCEDURE raise_error_with_params(IN name VARCHAR, OUT strlength INTEGER) BEGIN IF (name = 'xx') THEN SIGNAL SQLSTATE '20001' SET MESSAGE_TEXT = 'test exception'; END IF; SET strlength = LENGTH(name); END;
