+++ dbfit-java/oracle/src/main/java/dbfit/environment/OracleEnvironment.java
@@ -556,9 +556,9 @@ public class OracleEnvironment extends AbstractDbEnvironment {
     }
 
     //@Override
-    //public String getActualErrorCode(SQLException e) {
-//System.out.println("in Oracle getActualErrorCode");
-        //return Integer.toString(e.getErrorCode());
-  //      return e.getSQLState();
-    //}
+    public String getActualErrorCode(SQLException e) {
+System.out.println("in Oracle getActualErrorCode");
+        return Integer.toString(e.getErrorCode());
+        //return e.getSQLState();
+    }
 }
