+++ test_vm/scripts/oracle/sql/acceptancescripts-Oracle.sql
@@ -141,16 +141,18 @@ create or replace public synonym pub_syn_animals for dfsyntest.animals;
 
 create or replace procedure raise_error_with_params(name in varchar2, strlength out number) as
 begin
-    if (param1 = 'xx') then
+    if (name = 'xx') then
         raise_application_error(-20001,'test exception');
     end if;
     strlength := LENGTH(name);
 end;
+/
 
 create or replace procedure raise_error_no_params as
 begin
     raise_application_error(-20001,'test exception');
 end;
+/
 
 exit
